KEGCL-main/requirements.txt
astor	0.8.1
ca-certificates	2024.9.24
certifi	2022.12.7
charset-normalizer	3.4.0
cloudpickle	2.2.1
colorama	0.4.6
cycler	0.11.0
filelock	3.11.0
googledrivedownloader	0.4
idna	3.1
importlib-metadata	4.13.0
isodate	0.6.1
Jinja2	3.1.4
joblib	1.3.2
json-tricks	3.17.3
kiwisolver	1.4.5
littleutils	0.2.2
MarkupSafe	2.1.5
matplotlib	3.3.2
networkx	2.6.3
nni	2.10.1
numpy	1.21.6
ogb	1.3.6
openssl	3.0.15
outdated	0.2.2
packaging	24
pandas	1.3.5
Pillow	9.5.0
pip	22.3.1
prettytable	3.7.0
psutil	6.1.0
pyparsing	3.1.4
python	3.7.1
python-dateutil	2.9.0.post0
python-louvain	0.16
PythonWebHDFS	0.2.3
pytz	2024.2
PyYAML	6.0.1
rdflib	6.3.2
requests	2.31.0
responses	0.23.3
schema	0.7.7
scikit-learn	1.0.2
scipy	1.7.3
setuptools	65.6.3
simplejson	3.19.3
six	1.16.0
sqlite	3.45.3
threadpoolctl	3.1.0
torch	1.7.1
torch-cluster	1.5.9
torch-geometric	1.7.2
torch-scatter	2.0.6
torch-sparse	0.6.8
torch-spline-conv	1.2.1
tqdm	4.67.0
typeguard	2.13.3
types-PyYAML	6.0.12.12
typing_extensions	4.7.1
urllib3	1.26.10
vc	14.4
vs2015_runtime	14.40.33807
wcwidth	0.2.13
websockets	11.0.3
wheel	0.38.4
wincertstore	0.2
zipp	3.15.0
